Rename right menu store by mobile menu store
