+++ resources/css/app.css
@@ -49,23 +49,23 @@ p {
 
 .bottom_right_menu {
     background: url('https://content.demiplane.com/nexus/vampire/vtm-marble-red.jpg') !important;
-    opacity: 0;
+    opacity: 1;
     animation: fadeUp 300ms forwards;
 }
 
 .character_card {
     background: url('https://content.demiplane.com/nexus/vampire/vtm-marble-red.jpg') !important;
     @apply mx-4 px-4 py-2 rounded-lg border border-red-900;
+}
 
-    h2 {
-        font-family: 'Ubuntu', sans-serif;
-        @apply text-blood-500 font-bold text-2xl my-2;
-    }
+.character_card h2 {
+    font-family: 'Ubuntu', sans-serif;
+    @apply text-blood-500 font-bold text-2xl my-2;
+}
 
-    h3 {
-        font-family: 'Ubuntu', sans-serif;
-        @apply text-skin-50 text-xl my-2;
-    }
+.character_card h3 {
+    font-family: 'Ubuntu', sans-serif;
+    @apply text-skin-50 text-xl my-2;
 }
 
 .menu_text {
@@ -141,7 +141,7 @@ p {
     @apply my-2 ;
 }
 
-.mobile-menu{
+.mobile-menu {
     font-family: 'Ubuntu', sans-serif;
     @apply mt-4 flex justify-center italic font-bold text-white text-xl px-4 py-2 rounded-lg bg-blood-500 w-[180px] items-center;
 }
